requirements.txt
flask
flask-sqlalchemy
flask-cors
pyjwt
gunicorn
